improve install script

+++ release/install.sh
@@ -3,9 +3,11 @@ set -e
 
 version=$1
 
-[ -z "$version" ] && { echo "please provide a version number"; exit 1; }
-
 lib_name=posix_shell_lib
+repo="j-khong/$lib_name"
+last_version=$(curl -s "https://api.github.com/repos/$repo/tags" | grep '"name"' | head -n 1 | awk -F '"' '{print $4}')
+
+[ -z "$version" ] && { version=$last_version; }
 
 install_folder_name=".$lib_name"
 root_install_folder=~/$install_folder_name
@@ -16,13 +18,12 @@ if [ -d "$installed_version_folder" ]; then
     exit 0
 fi
 
-URL="https://github.com/j-khong/${lib_name}/releases/download/${version}/${lib_name}-${version}.tar.gz"
+URL="https://github.com/${repo}/releases/download/${version}/${lib_name}-${version}.tar.gz"
 
 if curl -o /dev/null -s -w "%{http_code}" -L "$URL" | grep -q "^2"; then
     :
 else
     echo "❌ the version you provided [$version] does not exist"
-    echo $URL
     exit 1
 fi
 
@@ -33,15 +34,19 @@ echo "📦 installing package"
 tar -xzf ${lib_name}.tar.gz -C "$root_install_folder"
 rm ${lib_name}.tar.gz
 
-
-symlink_name="$root_install_folder/latest"
-if [ -L "$symlink_name" ]; then rm "$symlink_name"; fi
-ln -s "$installed_version_folder" "$symlink_name"
+if [ "$version" == "$last_version" ]; then
+    symlink_name="$root_install_folder/latest"
+    if [ -L "$symlink_name" ]; then rm "$symlink_name"; fi
+    ln -s "$installed_version_folder" "$symlink_name"
+fi
 
 echo "✅ installation done !"
 
 echo ""
 
 echo "you can import a specific lib file in your scripts with :"
-echo ". $HOME/$install_folder_name/$version/{the_file_path} or "
-echo ". $HOME/$install_folder_name/latest/{the_file_path}"
+echo ". $HOME/$install_folder_name/$version/{the_file_path}"
+if [ "$version" == "$last_version" ]; then
+    echo "or"
+    echo ". $HOME/$install_folder_name/latest/{the_file_path}"
+fi
